Distorted tree rings
From https://www.earth.columbia.edu/news/2004/story06-03-04.html
Lamont dendrochronologist Gordon Jacoby collected this sample of tree rings from a Siberian pine in Mongolia, which records the years 534-539 C.E. (left to right). The narrow, distorted rings for 536 and 537 indicate a drastic cooling in the northern hemisphere that froze sap in the cells during the growing season. Evidence for this abrupt climate change points to a massive eruption of the volcanic precursor of Krakatoa. Another theory invokes a cosmic impact that possibly triggered the eruption. Photo by Dee Breger